Added new CAN emssage and tested on all systems.

+++ rm48_fsw/user_include/can_comms.h
@@ -13,7 +13,7 @@ typedef enum {
 void get_sensor_msg(uint8_t* msg_data, sensor_msg m);
 float get_last_can_height_msg(void);
 float get_last_can_heading_msg(void);
-void get_last_attitude(int16_t *att);
+void get_last_attitude(float *att);
 
 void publish_roll_pitch(float roll, float pitch);
 void publish_yaw_height_estimate(float yaw, float height_estimate);
